WIP: Bump upstreams for Silicon
Adopt the following upstream versions:
- odlparent-8.0.0
- yangtools-6.0.0
- mdsal-7.0.0
- controller-3.0.0
- infrautils-1.9.0

Since we are forced to migrate some of the code, perform related
migrations if they are trivial to do.

Quite a few tests are missing proper keys and properties, hence
these are added in the most obvious manner possible. If there is
a list/map warning in that vicinity, also use BindingMap.of() to get
rid of the warning.

Also a few obvious modernizer warnings are fixed, as it's easier
than add suppressions.

+++ openflowplugin-impl/src/test/java/org/opendaylight/openflowplugin/impl/OpenFlowPluginProviderImplTest.java
@@ -89,7 +89,7 @@ public class OpenFlowPluginProviderImplTest {
     private static final int RPC_REQUESTS_QUOTA = 500;
     private static final long GLOBAL_NOTIFICATION_QUOTA = 131072;
     private static final Uint16 THREAD_POOL_MIN_THREADS = Uint16.ONE;
-    private static final int THREAD_POOL_MAX_THREADS = 32000;
+    private static final Uint16 THREAD_POOL_MAX_THREADS = Uint16.valueOf(32000);
     private static final Uint32 THREAD_POOL_TIMEOUT = Uint32.valueOf(60);
     private static final long BASIC_TIMER_DELAY = 1L;
     private static final boolean USE_SINGLE_LAYER_SERIALIZATION = false;
